+++ src/Rfc4122/UuidV1.php
@@ -22,6 +22,7 @@ use Ramsey\Uuid\Converter\TimeConverterInterface;
 use Ramsey\Uuid\Exception\DateTimeException;
 use Ramsey\Uuid\Exception\InvalidArgumentException;
 use Ramsey\Uuid\Rfc4122\FieldsInterface as Rfc4122FieldsInterface;
+use Ramsey\Uuid\TimeBasedInterface;
 use Ramsey\Uuid\Uuid;
 use Throwable;
 
@@ -35,7 +36,7 @@ use const STR_PAD_LEFT;
  *
  * @psalm-immutable
  */
-final class UuidV1 extends Uuid implements UuidInterface
+final class UuidV1 extends Uuid implements UuidInterface, TimeBasedInterface
 {
     /**
      * Creates a version 1 (time-based) UUID
@@ -64,16 +65,6 @@ final class UuidV1 extends Uuid implements UuidInterface
         parent::__construct($fields, $numberConverter, $codec, $timeConverter);
     }
 
-    /**
-     * Returns a DateTimeInterface object representing the timestamp associated
-     * with the UUID
-     *
-     * The timestamp value is only meaningful in a time-based UUID, which
-     * has version type 1.
-     *
-     * @return DateTimeImmutable A PHP DateTimeImmutable instance representing
-     *     the timestamp of a version 1 UUID
-     */
     public function getDateTime(): DateTimeInterface
     {
         $time = $this->timeConverter->convertTime($this->fields->getTimestamp());