+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+* Introduce `Ramsey\Uuid\TimeBasedInterface` to represent any UUID based on a
+  date/time value.
+
 ### Changed
 
 * Remove the following deprecated classes: